to: makes model download when project start and download to local folder
