+++ backend/src/main.ts
@@ -1,6 +1,7 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import 'reflect-metadata';
+import { downloadAllModels } from './model/ModelLoader';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
@@ -16,6 +17,8 @@ async function bootstrap() {
       'Access-Control-Allow-Credentials',
     ],
   });
+  await downloadAllModels();
   await app.listen(process.env.PORT ?? 3000);
 }
+
 bootstrap();
